requirements.txt
fastapi
uvicorn
bcrypt
pydantic
sqlalchemy
pyodbc
pytest  
python-jose
python-multipart